fmt and restructure tests
