+++ interp/src/environment.rs
@@ -85,7 +85,7 @@ impl InterpreterState {
             "std_slice" => Box::new(combinational::StdSlice::new(params)),
             "std_pad" => Box::new(combinational::StdPad::new(params)),
             "std_reg" => Box::new(stateful::StdReg::new(params)),
-            "std_const" => Box::new(stateful::StdConst::new(params)),
+            "std_const" => Box::new(combinational::StdConst::new(params)),
             "std_mem_d1" => {
                 let mut prim = Box::new(stateful::StdMemD1::new(params));
 
@@ -173,236 +173,6 @@ impl InterpreterState {
         }
         Rc::new(RefCell::new(map))
     }
-
-    //all of these use parameters as values for constuctors
-    /* fn construct_cp_map(ctx: &ir::Context) -> PrimitiveMap {
-        let mut map = HashMap::new();
-        for comp in &ctx.components {
-            for cell in comp.cells.iter() {
-                let cl: &ir::Cell = &cell.borrow();
-
-                if let Some(name) = &cl.type_name() {
-                    match name.as_ref() {
-                        "std_reg" => {
-                            let reg = primitives::StdReg::new(
-                                cl.get_parameter("WIDTH").unwrap(),
-                            );
-                            map.insert(cl as ConstCell, Primitive::StdReg(reg));
-                        }
-                        "std_const" => {
-                            let width = cl.get_parameter("WIDTH").unwrap();
-                            let cst = primitives::StdConst::new(
-                                width,
-                                Value::try_from_init(
-                                    cl.get_parameter("VALUE").unwrap(),
-                                    width,
-                                )
-                                .unwrap(),
-                            );
-                            map.insert(
-                                cl as ConstCell,
-                                Primitive::StdConst(cst),
-                            );
-                        }
-                        "std_lsh" => {
-                            let width = cl.get_parameter("WIDTH").unwrap();
-                            let lsh = primitives::StdLsh::new(width);
-                            map.insert(cl as ConstCell, Primitive::StdLsh(lsh));
-                        }
-                        "std_rsh" => {
-                            let width = cl.get_parameter("WIDTH").unwrap();
-                            let rsh = primitives::StdRsh::new(width);
-                            map.insert(cl as ConstCell, Primitive::StdRsh(rsh));
-                        }
-                        "std_add" => {
-                            let adder = primitives::StdAdd::new(
-                                cl.get_parameter("WIDTH").unwrap(),
-                            );
-                            map.insert(
-                                cl as ConstCell,
-                                Primitive::StdAdd(adder),
-                            );
-                        }
-                        "std_sub" => {
-                            let width = cl.get_parameter("WIDTH").unwrap();
-                            let sub = primitives::StdSub::new(width);
-                            map.insert(cl as ConstCell, Primitive::StdSub(sub));
-                        }
-                        "std_slice" => {
-                            let slc = primitives::StdSlice::new(
-                                cl.get_parameter("IN_WIDTH").unwrap(),
-                                cl.get_parameter("OUT_WIDTH").unwrap(),
-                            );
-                            map.insert(
-                                cl as ConstCell,
-                                Primitive::StdSlice(slc),
-                            );
-                        }
-                        "std_pad" => {
-                            let pad = primitives::StdPad::new(
-                                cl.get_parameter("IN_WIDTH").unwrap(),
-                                cl.get_parameter("OUT_WIDTH").unwrap(),
-                            );
-                            map.insert(cl as ConstCell, Primitive::StdPad(pad));
-                        }
-                        "std_not" => {
-                            let not = primitives::StdNot::new(
-                                cl.get_parameter("WIDTH").unwrap(),
-                            );
-                            map.insert(cl as ConstCell, Primitive::StdNot(not));
-                        }
-                        "std_and" => {
-                            let and = primitives::StdAnd::new(
-                                cl.get_parameter("WIDTH").unwrap(),
-                            );
-                            map.insert(cl as ConstCell, Primitive::StdAnd(and));
-                        }
-                        "std_or" => {
-                            let or = primitives::StdOr::new(
-                                cl.get_parameter("WIDTH").unwrap(),
-                            );
-                            map.insert(cl as ConstCell, Primitive::StdOr(or));
-                        }
-                        "std_xor" => {
-                            let xor = primitives::StdXor::new(
-                                cl.get_parameter("WIDTH").unwrap(),
-                            );
-                            map.insert(cl as ConstCell, Primitive::StdXor(xor));
-                        }
-                        "std_ge" => {
-                            let ge = primitives::StdGe::new(
-                                cl.get_parameter("WIDTH").unwrap(),
-                            );
-                            map.insert(cl as ConstCell, Primitive::StdGe(ge));
-                        }
-                        "std_gt" => {
-                            let gt = primitives::StdGt::new(
-                                cl.get_parameter("WIDTH").unwrap(),
-                            );
-                            map.insert(cl as ConstCell, Primitive::StdGt(gt));
-                        }
-                        "std_eq" => {
-                            let eq = primitives::StdEq::new(
-                                cl.get_parameter("WIDTH").unwrap(),
-                            );
-                            map.insert(cl as ConstCell, Primitive::StdEq(eq));
-                        }
-                        "std_neq" => {
-                            let neq = primitives::StdNeq::new(
-                                cl.get_parameter("WIDTH").unwrap(),
-                            );
-                            map.insert(cl as ConstCell, Primitive::StdNeq(neq));
-                        }
-                        "std_le" => {
-                            let le = primitives::StdLe::new(
-                                cl.get_parameter("WIDTH").unwrap(),
-                            );
-                            map.insert(cl as ConstCell, Primitive::StdLe(le));
-                        }
-                        "std_lt" => {
-                            let lt = primitives::StdLt::new(
-                                cl.get_parameter("WIDTH").unwrap(),
-                            );
-                            map.insert(cl as ConstCell, Primitive::StdLt(lt));
-                        }
-                        "std_mem_d1" => {
-                            let mut m1 = primitives::StdMemD1::new(
-                                cl.get_parameter("WIDTH").unwrap(),
-                                cl.get_parameter("SIZE").unwrap(),
-                                cl.get_parameter("IDX_SIZE").unwrap(),
-                            );
-                            if let Some(v) = mems
-                                .as_ref()
-                                .map(|x| x.get(cell_name))
-                                .flatten()
-                            {
-                                m1.initialize_memory(v)
-                            }
-
-                            map.insert(
-                                cl as ConstCell,
-                                Primitive::StdMemD1(m1),
-                            );
-                        }
-                        "std_mem_d2" => {
-                            let mut m2 = primitives::StdMemD2::new(
-                                cl.get_parameter("WIDTH").unwrap(),
-                                cl.get_parameter("D0_SIZE").unwrap(),
-                                cl.get_parameter("D1_SIZE").unwrap(),
-                                cl.get_parameter("D0_IDX_SIZE").unwrap(),
-                                cl.get_parameter("D1_IDX_SIZE").unwrap(),
-                            );
-
-                            if let Some(v) = mems
-                                .as_ref()
-                                .map(|x| x.get(cell_name))
-                                .flatten()
-                            {
-                                m2.initialize_memory(v)
-                            }
-
-                            map.insert(
-                                cl as ConstCell,
-                                Primitive::StdMemD2(m2),
-                            );
-                        }
-                        "std_mem_d3" => {
-                            let mut m3 = primitives::StdMemD3::new(
-                                cl.get_parameter("WIDTH").unwrap(),
-                                cl.get_parameter("D0_SIZE").unwrap(),
-                                cl.get_parameter("D1_SIZE").unwrap(),
-                                cl.get_parameter("D2_SIZE").unwrap(),
-                                cl.get_parameter("D0_IDX_SIZE").unwrap(),
-                                cl.get_parameter("D1_IDX_SIZE").unwrap(),
-                                cl.get_parameter("D2_IDX_SIZE").unwrap(),
-                            );
-
-                            if let Some(v) = mems
-                                .as_ref()
-                                .map(|x| x.get(cell_name))
-                                .flatten()
-                            {
-                                m3.initialize_memory(v)
-                            }
-
-                            map.insert(
-                                cl as ConstCell,
-                                Primitive::StdMemD3(m3),
-                            );
-                        }
-                        "std_mem_d4" => {
-                            let mut m4 = primitives::StdMemD4::new(
-                                cl.get_parameter("WIDTH").unwrap(),
-                                cl.get_parameter("D0_SIZE").unwrap(),
-                                cl.get_parameter("D1_SIZE").unwrap(),
-                                cl.get_parameter("D2_SIZE").unwrap(),
-                                cl.get_parameter("D3_SIZE").unwrap(),
-                                cl.get_parameter("D0_IDX_SIZE").unwrap(),
-                                cl.get_parameter("D1_IDX_SIZE").unwrap(),
-                                cl.get_parameter("D2_IDX_SIZE").unwrap(),
-                                cl.get_parameter("D3_IDX_SIZE").unwrap(),
-                            );
-
-                            if let Some(v) = mems
-                                .as_ref()
-                                .map(|x| x.get(cell_name))
-                                .flatten()
-                            {
-                                m4.initialize_memory(v)
-                            }
-
-                            map.insert(
-                                cl as ConstCell,
-                                Primitive::StdMemD4(m4),
-                            );
-                        }
-                        e => panic!("Unknown primitive {}", e),
-                    }
-                }
-            }
-        }
-        Rc::new(RefCell::new(map))
-    } */
 
     fn construct_pv_map(ctx: &ir::Context) -> PortValMap {
         let mut map = HashMap::new();
