+++ interp/src/primitives/combinational.rs
@@ -1,7 +1,61 @@
-use super::Primitive;
+use super::{stateful::get_param, Primitive};
 use crate::comb_primitive;
 use crate::values::Value;
+use calyx::ir;
 use std::ops::Not;
+
+/// A constant.
+#[derive(Default, Debug)]
+pub struct StdConst {
+    value: Value,
+}
+
+impl StdConst {
+    pub fn from_constants(value: u64, width: u64) -> Self {
+        StdConst {
+            value: Value::try_from_init(value, width).unwrap(),
+        }
+    }
+
+    pub fn new(params: ir::Binding) -> Self {
+        let width = get_param(&params, "WIDTH")
+            .expect("Missing width parameter from std_const binding");
+
+        let init_value = get_param(&params, "VALUE")
+            .expect("Missing `vale` param from std_const binding");
+
+        let value = Value::try_from_init(init_value, width).unwrap();
+
+        Self { value }
+    }
+}
+
+impl Primitive for StdConst {
+    fn is_comb(&self) -> bool {
+        true
+    }
+
+    fn validate(&self, _inputs: &[(ir::Id, &Value)]) {}
+
+    fn execute(
+        &mut self,
+        _inputs: &[(ir::Id, &Value)],
+        _done_val: Option<&Value>,
+    ) -> Vec<(ir::Id, crate::values::OutputValue)> {
+        vec![("out".into(), self.value.clone().into())]
+    }
+
+    fn reset(
+        &mut self,
+        _inputs: &[(ir::Id, &Value)],
+    ) -> Vec<(ir::Id, crate::values::OutputValue)> {
+        vec![("out".into(), self.value.clone().into())]
+    }
+
+    fn commit_updates(&mut self) {}
+
+    fn clear_update_buffer(&mut self) {}
+}
 
 // ===================== Unary operations ======================
 comb_primitive!(StdNot[WIDTH](r#in: WIDTH) -> (out: WIDTH) {
