+++ interp/src/tests/mod.rs
@@ -1,0 +1,3 @@
+mod primitives;
+mod stk_env;
+mod values;
