+++ interp/src/tests/values.rs
@@ -1,0 +1,39 @@
+#[cfg(test)]
+mod val_test {
+    use crate::values::Value;
+    #[test]
+    fn basic_print_test() {
+        let v1 = Value::try_from_init(12, 5).unwrap();
+        println!("12 with bit width 5: {}", v1);
+        assert_eq!(v1.as_u64(), 12);
+    }
+    #[test]
+    fn basic_print_test2() {
+        let v1 = Value::try_from_init(33, 6).unwrap();
+        println!("33 with bit width 6: {}", v1);
+        assert_eq!(v1.as_u64(), 33);
+    }
+    #[test]
+    fn too_few_bits() {
+        let v_16_4 = Value::try_from_init(16, 4).unwrap();
+        println!("16 with bit width 4: {}", v_16_4);
+        assert_eq!(v_16_4.as_u64(), 0);
+        let v_31_4 = Value::try_from_init(31, 4).unwrap();
+        println!("31 with bit width 4: {}", v_31_4);
+        let v_15_4 = Value::try_from_init(15, 4).unwrap();
+        println!("15 with bit width 4: {}", v_15_4);
+        assert_eq!(v_31_4.as_u64(), v_15_4.as_u64());
+    }
+    #[test]
+    fn clear() {
+        let v_15_4 = Value::try_from_init(15, 4).unwrap();
+        let v_15_4 = v_15_4.clear();
+        println!("15 with bit width 4 AFTER clear: {}", v_15_4);
+        assert_eq!(v_15_4.as_u64(), 0);
+    }
+    #[test]
+    fn ext() {
+        let v_15_4 = Value::try_from_init(15, 4).unwrap();
+        assert_eq!(v_15_4.as_u64(), v_15_4.ext(8).as_u64());
+    }
+}
